chore: slight reorganization
